securing the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.views.generic import CreateView
 from .models import Product, Review
@@ -37,8 +38,13 @@ class AddProductReview(CreateView):
     template_name = 'products/add_review.html'
 
 
+@login_required
 def add_product(request):
     
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only superusers can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -58,7 +64,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only superusers can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Successfully deleted item from ERR Store.')
